+++ lib/lightning/services/usage_limiter.ex
@@ -26,5 +26,15 @@ defmodule Lightning.Services.UsageLimiter do
     adapter().get_run_options(context)
   end
 
+  @impl true
+  def get_data_retention_periods(context) do
+    adapter().get_data_retention_periods(context)
+  end
+
+  @impl true
+  def get_data_retention_message(context) do
+    adapter().get_data_retention_message(context)
+  end
+
   defp adapter, do: adapter(:usage_limiter)
 end