+++ lib/lightning_web/live/project_live/new_collaborator_component.ex
@@ -4,7 +4,7 @@ defmodule LightningWeb.ProjectLive.NewCollaboratorComponent do
   use LightningWeb, :live_component
 
   alias Lightning.Projects
-  alias Lightning.Projects.ProjectUsersLimiter
+  alias Lightning.Projects.ProjectLimiter
   alias LightningWeb.ProjectLive.Collaborators
   alias Phoenix.LiveView.JS
 
@@ -130,7 +130,7 @@ defmodule LightningWeb.ProjectLive.NewCollaboratorComponent do
 
     project_users = Ecto.Changeset.get_embed(changeset, :collaborators)
 
-    case ProjectUsersLimiter.request_new(
+    case ProjectLimiter.request_new_user(
            socket.assigns.project.id,
            Enum.count(project_users)
          ) do