+++ test/lightning_web/live/project_live_test.exs
@@ -1852,6 +1852,71 @@ defmodule LightningWeb.ProjectLiveTest do
     end
 
     @tag role: :admin
+    test "history retention period only shows provided options by the extension",
+         %{
+           conn: conn,
+           project: %{id: project_id} = project
+         } do
+      expected_options = [7, 14]
+      other_options = [30, 90, 180, 365]
+
+      expect(
+        Lightning.Extensions.MockUsageLimiter,
+        :get_data_retention_periods,
+        2,
+        fn %{project_id: ^project_id} ->
+          expected_options
+        end
+      )
+
+      {:ok, view, _html} =
+        live(
+          conn,
+          ~p"/projects/#{project.id}/settings#data-storage"
+        )
+
+      html =
+        view
+        |> element("select#retention-settings-form_history_retention_period")
+        |> render()
+
+      for option <- expected_options do
+        assert html =~ "#{option} Days</option>"
+      end
+
+      for option <- other_options do
+        refute html =~ "#{option} Days</option>"
+      end
+    end
+
+    @tag role: :admin
+    test "history retention period is disabled when only one option is provided",
+         %{
+           conn: conn,
+           project: %{id: project_id} = project
+         } do
+      expect(
+        Lightning.Extensions.MockUsageLimiter,
+        :get_data_retention_periods,
+        2,
+        fn %{project_id: ^project_id} ->
+          [7]
+        end
+      )
+
+      {:ok, view, _html} =
+        live(
+          conn,
+          ~p"/projects/#{project.id}/settings#data-storage"
+        )
+
+      assert has_element?(
+               view,
+               "#retention-settings-form_history_retention_period:disabled"
+             )
+    end
+
+    @tag role: :admin
     test "dataclip retention period is disabled if the history period has not been set",
          %{
            conn: conn,