+++ test/support/conn_case.ex
@@ -57,6 +57,11 @@ defmodule LightningWeb.ConnCase do
       Lightning.Extensions.AccountHook
     )
 
+    Mox.stub_with(
+      Lightning.Extensions.MockProjectHook,
+      Lightning.Extensions.ProjectHook
+    )
+
     pid =
       Ecto.Adapters.SQL.Sandbox.start_owner!(Lightning.Repo,
         shared: not tags[:async]