+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to
 - Additional documentation and notification text relating to the importance of
   alternate storage for Kafka triggers.
   [#2614](https://github.com/OpenFn/lightning/issues/2614)
+- Add support for run memory limit option
+  [#2623](https://github.com/OpenFn/lightning/pull/2623)
 
 ### Changed